Add basic Docker setup
This setup runs two containers, one for WordPress and one for the
database. A third container is used to run the WordPress CLI for setting
up the environment.

The environment can be spun up with a standard `docker-compose up` and
then a script is provided to perform the rest of the development
environment setup.