Create a Stripe account ID setting
We'll need this until the actual authentication is implemented and then
we can defer the fetching of the account ID to the payments API.

It would be nicer to inject the account ID to use into the API client
at the point the client is created. However, we don't have access to the
setting at that point so we set it up here instead.

We also pass the account ID out to Stripe.js so that the payment form
uses the correct account.