Retrieve data to display in TransactionsList using transactions selector
By retrieving the relevant selectors we can access the data from the
fresh-data data store, and use that to populate the transactions table.

As far as I understand this (@coderkevin would be a better person to
ask) the `withSelect` method basically injects data from the selector
into the component's `this.props`, allowing us to use the data in the
component.