Get WooCommerce with Composer
This allows us to run our unit tests with no further setup required. We
always need to run `composer install` to get PHPUnit and other
development dependencies, so it's not more steps to pull down
WooCommerce as well.

The WooCommerce package uses `composer/installers` to change its
install path to something that makes sense for WordPress installations.
We disable this with `installer-disable` option so that WooCommerce is
installed to the vendor folder along with all our other development time
packages.