Display empty table message using the summary property of 'TableCard'
As discussed in #97 using the summary property instead of an
'EmptyTable' component looks more consistent than using just the
'EmptyTable'.

This has the caveat of using the summary for two purposes: displaying a
message when there is no data available to display, and then displaying
an actual summary when there _is_ data to display.
This requires some boolean logic that I tried to make expressive so it's
easy to follow.