Cache result when reading plugin headers
In order to avoid multiple calls to get_file_data and consequently fopen, we've
encapsulated the get plugin header functionality in a method that caches the
result of the first call. Therefore, we're able to get the headers in multiple
places without worrying about the overhead of opening the same file multiple times.