Add plugin version to enqueue scripts and styles based on SCRIPT_DEBUG (#190)
* Add version number constant to woocommerce-payments

* Base scripts/styles version number on SCRIPT_DEBUG

Similar to woocommerce-admin, use the constant SCRIPT_DEBUG to decide
the version number when enqueuing scripts/styles.

If running with SCRIPT_DEBUG false, the version of imports will be the
same as the current WCPay version. If true, the version will be the
timestamp for the file being imported.

* Remove static version number

Version number will be extracted from the plugin headers in order to
decrase the number of places it needs to be maintained.

* Cache result when reading plugin headers

In order to avoid multiple calls to get_file_data and consequently fopen, we've
encapsulated the get plugin header functionality in a method that caches the
result of the first call. Therefore, we're able to get the headers in multiple
places without worrying about the overhead of opening the same file multiple times.

* Set version number dynamically in WC_Payments init

Set version number based on the value for Version in the header of
woocommerce-payments.php.