Move mocked data from client frontend to client API
Instead of mocking data in the selectors it is now mocked in the client
API. These mocks will eventually be moved to the Payments Server, and
then switched out for a live version there.

Tests are changed to reflect the recent updates.

This commit also removes selectors that are no longer being used, in
favor of ones that use the pagination.