Fix transactions table re-rendering issue
Transactions table wasn't rendering due to the fact that the props being
passed to the component didn't change. Because of this the table didn't
re-render when incoming data requests were resolved.

The solution is to use the URL query parameters to control the table.
This gives us two benefits:
  1. The table re-renders (yay!).
  2. The table pages are added to the browser's history allowing you to
     use the browser's back/forward functions.

There's still some work to be done to properly implement the
back/forward functionality, but at least the table is working now!