Fix update of extracted settings when form is submitted
When submitting the save settings form, the settings values that were extracted
to the class instance were not updated.

Moved the extraction code to a function and hooked it to
woocommerce_update_options_payment_gateways_{id}, as well as called it on the
constructor for when the page is loaded with a GET request.