Pass account data to check requirements
The account object contains data regarding the requirements, such as current_deadline, which needs to be used outside of the account_has_pending_requirements method.

To avoid calling the account API multiple times, account became a parameter of that method and the caller is responsible for actually fetching the account data.