Update functional prototype settings (#211)
* Remove Title, Description, Stripe account and publishable key settings

* Move test mode and enable/disable checkbox to the bottom of the form

* Update Description setting default value

* Conditionally display Stripe connect or dashboard link

Display connect to a new Stripe account link only when stripe_account_id
or the respective publishable_key are not set. If the account is already
connected, display the link to the Express Dashboard.

* Rename Payment Details label to Account and change link text

* Update Stripe account disconnected message

* Add getters for test_mode and publishable_key

Remove extract_settings_values and encapsulate test_mode and publishable_key
attributes with getters. This should help with when the settings are updated
and removes the need to update the class attributes when the settings change.