Fix settings sub-navigation disappearance
Add a temporary fix to the settings menu, which was conflicting with the /transactions/details page. Issue #231 will be kept open until we develop a more robust way to fix this problem.

In order for the details page to be loaded when navigating directly to its url, it needs to be registered in wc-admin. However, registering a page also creates a menu entry for it.

In order to create the transaction details page we registered a new /transaction/details page. We don't want to have a menu entry for this page, so it was nested in a way that it was not rendered. This broke the settings menu because it was added manually to the last menu entry ($submenu[ $last_submenu_key ][] = array( ... );).

This solution to fix the settings menu is to register the details page only after we manually place the settings entry, but that does not fix the issue that we actually do not want any kind of menu entry for that page and we're hiding it in a hacky way.