Set PHP 7.0 as the default platform for Composer
Ideally we would set this to our minimum supported version of PHP (5.6
at the time of writing). This ensures that no matter what version of
PHP a developer might be using locally, Composer will only install
packages compatible with the configured PHP version.

This allows us to run `composer install` with any version of PHP greater
than 7.0. If we didn't do this then Composer might add packages only
supported by higher versions to our lock file, which then makes it
impossible to go back and use 7.0 (for another developer or CI system).

Unfortunately `woocommerce/woocommerce-sniffs` has a minimum version of
7.0, so we can't go below that without removing the package. As a
compromise the PHP config is set to 7.0 and we'll take any manual steps
necessary to get 5.6 working on a CI server.

We also downgrade PHPUnit to 5.7.27 since this is the newest version
that has PHP 5.6 support.