Add filter allowing modification of arguments on OAuth init server call
In order to force the on-boarding to re-run we need to pass the
`force_on_boarding` argument. We're using this filter to do it because
we don't want to add it to the client method's parameters and give the
impression that this is fully supported functionality.

The filter could be used like this:

function add_force_on_boarding_arg( $get_oauth_data_args ) {
	$get_oauth_data_args['force_on_boarding'] = true;
	return $get_oauth_data_args;
}

add_filter(
	'wc_payments_get_oauth_data_args',
	'add_force_on_boarding_arg'
);