Get WooCommerce with Composer (#166)
* Get WooCommerce with Composer

This allows us to run our unit tests with no further setup required. We
always need to run `composer install` to get PHPUnit and other
development dependencies, so it's not more steps to pull down
WooCommerce as well.

The WooCommerce package uses `composer/installers` to change its
install path to something that makes sense for WordPress installations.
We disable this with `installer-disable` option so that WooCommerce is
installed to the vendor folder along with all our other development time
packages.

* Load WooCommerce from vendor folder

Instead of requiring a side by side install or WooCommerce we can now
rely on it being present in the vendor folder. Let's load it from there
instead.

* Ignore vendor folder when running JS tests

The vendor folder now contains JavaScript unit tests related to
WooCommerce. We don't want to run these as part of our testing.

* Set PHP 7.0 as the default platform for Composer

Ideally we would set this to our minimum supported version of PHP (5.6
at the time of writing). This ensures that no matter what version of
PHP a developer might be using locally, Composer will only install
packages compatible with the configured PHP version.

This allows us to run `composer install` with any version of PHP greater
than 7.0. If we didn't do this then Composer might add packages only
supported by higher versions to our lock file, which then makes it
impossible to go back and use 7.0 (for another developer or CI system).

Unfortunately `woocommerce/woocommerce-sniffs` has a minimum version of
7.0, so we can't go below that without removing the package. As a
compromise the PHP config is set to 7.0 and we'll take any manual steps
necessary to get 5.6 working on a CI server.

We also downgrade PHPUnit to 5.7.27 since this is the newest version
that has PHP 5.6 support.

* Remove special cases for CI setup and running unit tests

Now that dependencies are handled by Composer consistently across PHP
versions we can remove this code from our CI setup and test runner
scripts.

* Remove manual WooCommerce installation

This is now handled by Composer. We also specify the WooCommerce version
in `composer.json`, so we can remove the WC_VERSION variable from the
scripts.

* Add note about potential PHP 5.6 issues