Turn on logging for the FTP mirroring process
The Travis build is timing out because nothing is being output:
```
No output has been received in the last 10m0s, this potentially
indicates a stalled build or something wrong with the build itself.
```
This causes the build to terminate. The FTP process outputting progress
should prevent this (assuming it isn't too much output to trip the 4MB
Travis log limit).