Update payment detail action button to point to order page
The previous design could be confusing. For example, clicking the "refund" action button would have directed a user to the order detail screen and from there they would have had to click the refund button to issue the refund. The link to the order page was removed from the details row.