Add test mode notices via a higher order component
The higher order component takes in a parameter that's used
to decide what the notice should say. The idea is that maintaining
the higher order component will require less effort than manually
adding a notice to each component.

This way any custom logic can be handled in one location based on
the parameter passed to the component, instead of maintaining
custom logic for each individual page (component).

This commit also fixes any snapshots that changed for testing.