Add pagination to transactions list
This modifies the API on the client so that it accepts the `page` and
`pagesize` parameters and pass that on to the Payments Server.
That API is then used to retrieve the correct data for the requested
page and page size.

The biggest part of this commit is just adding the data handling. The
data is saved into the wp.data store with a key made up of the query and
a prefix.
For example, if the transactions are made for page 2, with a page size
of 50, those transactions will be stored in a JS object like so:

{
  'transactions:{"paged":"2","perPage":"50"}': [
    transaction1,
    transaction2,
    ...
  ]
}

To help with this there is 1 new constant added - containing the
available wp.data store prefixes - and a helper function `getResourceId`
to help create the proper wp.data store key, given a prefix and a query.

There's nothing magical going on with the transactions page itself;
because we're using the query itself to control when the data is loaded
we can use builtin `TableCard` functions to handle actually navigating
the pages and page size.