Add basic transactions list pagination (#343)
* Update tests to match new data format

* Update client code to consume updated data format from server

This commit updates the client API and front-end code to fully accept
the "finalized" version of the transactions server endpoint. We now have
the previously missing information and can properly populate the
transactions table.

* Add pagination to transactions list

This modifies the API on the client so that it accepts the `page` and
`pagesize` parameters and pass that on to the Payments Server.
That API is then used to retrieve the correct data for the requested
page and page size.

The biggest part of this commit is just adding the data handling. The
data is saved into the wp.data store with a key made up of the query and
a prefix.
For example, if the transactions are made for page 2, with a page size
of 50, those transactions will be stored in a JS object like so:

{
  'transactions:{"paged":"2","perPage":"50"}': [
    transaction1,
    transaction2,
    ...
  ]
}

To help with this there is 1 new constant added - containing the
available wp.data store prefixes - and a helper function `getResourceId`
to help create the proper wp.data store key, given a prefix and a query.

There's nothing magical going on with the transactions page itself;
because we're using the query itself to control when the data is loaded
we can use builtin `TableCard` functions to handle actually navigating
the pages and page size.

* Associate transactions list errors with a query

Now each individual query can have it's own error, instead of before
where errors would be associated with all queries, i.e. only one error
could be stored at a time.

* Make transactions selectors a bit more robust

* Add simple test for `getResourceId` utility function