Display error banner on WP Admin if Jetpack is not connected
Stop WCPayments plugin initialization when Jetpack is not connected.

When new instance of WC_Payments_Account class is created as part of
the WCPayments plugin initialization, an HTTP request is sent with Jetpack Client
which leads to the fatal error when Jetpack is not connected.
This is why we have to check if Jetpack is connected before creating an instance of WC_Payments_Account class.