Define Psalm config
We point at an autoloader (the autoloader is hacky junk, but might form
the basis of being able to run some unit tests without creating a
WordPress environment every time).

The config is set to be as strict as possible, with the exception of a
rule asking for return annotations on methods that don't return
anything. That seemed strange so I've disabled it.

We also enable a rule that checks for @throw annotations which does
a better job of flagging nested throw statements than PHPCS.