Add dispute evidence file upload field (#419)
* Add file upload.

Reuse method from WC_Payments_REST_Controller.
Move generic error message to the API client file upload handler.

* Files can be saved via the 'Save for later' button.
* Displays upload status text during and after upload.
* Shows the file name on success, otherwise displays an error
  message.
* Add file remove button

The button removes the file most recently uploaded, allowing people
to upload a different file, although removing the file first isn't
strictly necessary at this point.

The file isn't truly removed from the evidence until the user saves
or submits the evidence.

* Use pre-defined studio styles instead of redefining them

* Add custom ini file to wp docker container

There is a custom.ini file added to the `wordpress` container at the build
time but every time it's updated the container needs to be rebuilt.

Mapping `wc-payments-php.ini` into the container for easy access and
modifications at runtime.

* Handle evidence upload for files exceeding the max size

There are three limits available:
  - server limit specified in php.ini, managed by hosting provider;
  - site limit for multisite setup;
  - stripe limit for evidence upload, specified in docs.
Current commit handles the case when the server limit is exceeded.
An informative message is displayed to the user.

* Send empty metadata as an object to the client

When the API response object has an empty metadata field it is serialized as
an empty array on the client by default. This causes problems with using
metadata as key-value storage on the client.

* Update evidence upload errors handling

The WCPayments Server does not transfer reason of upload failure to
WCPayments Client at the moment. Sending generic error message to the
client as a workaround.

* Add tests for FileUploadControl

* Fix submitting untouched evidence form.

When previously saved evidence is being submitted without changes it
does not update status because saved evidence not sent to the server.

* Improve errors handling for large file uploads

When large file (8+ MB) is uploaded file params are empty and error
is not detected. Added extra check for such a case.

* Prevent evidence form submission while an upload is in progress

* Update read-only evidence form

Disable upload buttons and hide remove file icon (trash).

Co-authored-by: Paul Dechov <dechovsky@gmail.com>
Co-authored-by: Vasily Belolapotkov <belolapotkov.v@gmail.com>
Co-authored-by: Daniel Rey López <daniel.rey.lopez@gmail.com>