Include the autoloader
With the autoloader in place we can remove the now unnecessary separate
includes for each class file.

Instead of adding a new require statement (and one to our unit test
bootstrap) when creating a new class, we just need to run
`composer dumpautoload`. The creation of the autoloader is handled by
`composer install` as well for developers freshly cloning the project.