Update PHP versions supported (#587)
* Stop testing on PHP 5.6, it's now unsupported

We don't need the comment about composer install potentially failing on
PHP 5.6 either. We are setting the minimum PHP version to 7.0 in our
composer.json, so the install should always work.

* Upgrade to PHPUnit 6

With the new minimum version of PHP set to 7.0, we can use PHPUnit 6.
This is the newest version that still supports PHP 7.0.