Create or fetch a customer on the server when processing a payment
If a user has a customer ID stored in their user options, we update it
with their details form the current order and pass the ID into the
create payment intent method.

For users with no existing customer ID (or for non-logged in users) we
create a new customer and use the ID from that.