Add support to forward encoded arrays in query string
When an array is supplied in querystring, the blog_id substitution fails since there
are multiple square brackets (%5D and %5B) in the URL and only one argument is provided
to sprintf.

This can be fixed by applying the sprintf blog_id substition to the URL's path part and
merging the querystring after it.