Add a note about handling a Stripe JavaScript upgrade
There's no suggestion that Stripe will release a v4 of their JavaScript
any time soon, but this theoretical situation came up as part of testing
whether we play nicely with other plugins that also use stripe.js.

Given this might happen at an unknown point in the future it felt worth
while having the note close to the code.