On Pay for Order block payment UI during auth
On the Pay for Order page, authentication
happens after a page reload. This causes the payment form
to lose the credit card info previously entered. To avoid user
confusion, this form should be hidden and blocked from
being submitted.

After the payment is complete, the user is redirected. If there
was an error processing the authentication, the payment form is
unblocked.