Prevent plugin initialization if Jetpack 8.1 or earlier is installed
Due to changes in the autoloader, if Jetpack 8.1 or earlier is installed alongside WCPay, the whole site would crash. The crash would happen in the `\Jetpack\Config` initialization code.

I added a quick & dirty check to show an admin notice if Jetpack 8.1 or earlier is installed, and prevent the rest of the plugin from initializing. I didn't want to spend a lot of effort in this so it's not as smart as the main dependency checker (which shows a link to update the dependency, tells you which version of the dependency you have, etc). I just wanted something temporary that's better than the site crashing.

Jetpack 8.2 (which is already compatible with WCPay) was released in February 2020. I think we can remove this piece of code in a few months, maybe when that version is 1 year old? Anyway, I left a TODO comment, if someone stumbles upon it in the distant future it can be removed.

To test:
- Install Jetpack 8.1 without using this WCPay branch. Note that the site crashes.
- Install Jetpack 8.2. Everything works fine.
- Install Jetpack 8.1 again, now using this WCPay branch. You'll see a notice telling you to upgrade, and you won't see the `[$] Payments` menu at all.
- Install Jetpack 8.2, still using this WCPay branch. Everything should work fine.
- Disable Jetpack. Everything should still work fine.