Set charge id metadata before payment_complete()
This is so that the _charge_id metadata is saved in case
something goes wrong while marking the payment complete.
If something goes wrong, the store owner will still be able to see
the related charge ID in the order edit screen, and will be able to
use it for example to process a refund, etc.

This is the same fix as seen in
https://github.com/Automattic/woocommerce-payments/pull/695