Mock WC_Order in process payment success test
This change was made to allow testing of more of the
side effects for processing a payment in `process_payment`.

Instead of testing `process_payment` directly, a new function
is written `process_payment_for_order` to allow for mocking the
WC_Order object.

The following changes were made to the existing tests:
- To check order meta values '_intent_id', '_charge_id', '_intention_status',
instead of retrieving the meta value after calling
`process_payment_for_order`, check that the method `update_meta_data`.
- Test that the order note was correct by checking the correct
note value was passed to the method `add_order_note`.

The tests for the return value were unchanged.

A new test was added for checking that `payment_complete` is called.