Do not send billing details when adding cards using my account page
Since the my account page does not include the billing details form, it doesn't make sense to add
the billing_details object when creating a payment method. In some cases, it could lead to some
issues with the customer name being undefined when creating the setup intent. This commit adds
an option to not send the billing details and enables it for the my account form.