Add customer country to the transaction metadata
When a payment method is using the add_payment_method form, it doesn't include any billing details.
Because of this, charges using that card do not have them as well. To solve this, we're using the
metadata added to the transaction when the payments are processed. This commit adds the customer
country to the transactions metadata, so it can be used as a fall back for those cases.