Fix checkout when user creation is enabled
When there's no account created, the checkout form doesn't have the
wc-woocommerce_payments-payment-token-new radio input to determine
if a saved payment method was selected. This would cause an issue in
which, if the radio didn't exist, the checkout handler would assume
that a saved payment method was selected, resulting in "Payment method
not found" errors.

This commit improves the saved payment method detection, to ensure that
payment methods are created in case no saved method is selected.