Fix missing payment status for disputes (#738)
* Refactor PaymentStatusChip and add mapping file

Remove logic from PaymentStatusChip and let the charge status be resolved in the payment
summary, consolidating its behavior with DisputeStatusChip.

* Source disputed payment status from dispute status chip mappings

* Add test to ensure that disputed statuses take precedence over refunds

When a charge is disputed with an inquiry, the merchant might choose to refund the charge,
that will end up with an inquiry closed and refunded status. The payment status should
reflect the inquiry closed instead of fully refunded.