Add saved cards support (#758)
* Add support for saving cards at checkout

* Add support for saving cards from my account

* Detach payment method from customer when token is deleted

* Allow for additional changes when updating customer

* Update default payment method on Stripe when updated in WC

* Populate tokens based on payment methods

This commit uses the woocommerce_get_customer_payment_tokens hook to fetch payment
methods and create tokens for the ones that are not in WC already. It also uses
a transient to avoid multiple API calls.

* Fix undefined index code issue when detaching non-attached payment methods

* Add saved cards checkout support

* Allow adding of SCA/3DS saved cards using my account page

When adding payment methods using the my account page, the setup intent needs to be
authorized at the front end. This commit creates the setup intent using an AJAX call
and then uses the Stripe SDK to authorize it if needed.

* Fix overlay blink after setup intent confirmation

WC core calls block() when add_payment_form is submitted and since the form is already blocked, the overlay
blinks when the form is blocked again. This commit fixes that by overriding blockUI default ignoreIfBlocked
value to true. The default value is restored once the form is submitted.

* Handle setup intent confirmation using Stripe JS

Instead of confirming the setup intent on creation, let Stripe JS handle all possible cases,
including succeeded, which had to be handled separately from the others if the intent was
already confirmed.

* Clear cached payment methods when tokens are updated

* Do not send billing details when adding cards using my account page

Since the my account page does not include the billing details form, it doesn't make sense to add
the billing_details object when creating a payment method. In some cases, it could lead to some
issues with the customer name being undefined when creating the setup intent. This commit adds
an option to not send the billing details and enables it for the my account form.

* Fix checkout when user creation is enabled

When there's no account created, the checkout form doesn't have the
wc-woocommerce_payments-payment-token-new radio input to determine
if a saved payment method was selected. This would cause an issue in
which, if the radio didn't exist, the checkout handler would assume
that a saved payment method was selected, resulting in "Payment method
not found" errors.

This commit improves the saved payment method detection, to ensure that
payment methods are created in case no saved method is selected.

* Update payment method details when processing payments

Payment methods created using the my account page do not have billing details. This
commit uses the checkout form data to update the payment method used on checkout.

* Prevent saving or updating payment methods from crashing checkout

This commit wraps the saved cards code in the checkout flow in try/catch blocks
and logs the error without rethrowing if an exception is thrown, to avoid complete
checkout crashes.