Remove copy of payment methods from the parent order
Renewal orders were created with the payment_tokens from the parent order, so when changing
the token for a subscription, we'd need to also update the parent order with the new method.

Subscriptions already copies the data from the subscription order when creating a new one,
so we don't need this method unless we really want to use the parent order payment methods.

(cherry picked from commit 87d39ef87421158f158b309a1e827142776cdf2d)