Fix failed payment token update by saving token before order update
The woocommerce_subscriptions_changed_failing_payment_method_{gateway} hook is called when
the renewal order has its status changed. Therefore, we need to add the payment method to
the renewal order before completing the payment, otherwise, we cannot update the subscription
with the new payment details. This commit moves the token saving to before the order status
update and conciliates intents and setup intents to reuse checkout code.