Derive `is_using_saved_payment_method` from token
Instead of storing an explicit member variable that indicates whether
the payment used a saved payment method, make that decision based on
whether the `Payment_Information` object has a token or not, since
tokens only exist for saved payment methods.

This commit also moves a lot of the complexity of checking for errors
out of the static functions and into the constructor. That way we're
only throwing an exception when trying to instantiate a
`Payment_Information` object with invalid data.

Static functions now return an expected value or `null`.