Fix unit tests
WC_Payment_Gateway_WCPay_Process_Payment_Test::test_exception_thrown had to be
modified because the notice and status update for failing orders have been moved
to process_payment, which cannot accept a mock order. This commit changes the
mock_order to use WC_Helper_Order::create_order, which is a real WC order and
is saved to the database.