Add subscriber payment method changes (#832)
* Add support for trial subscriptions and payment method changes

* Ensure tokens are retrieved in the order they were added

If we use WC_Payment_Tokens::get_order_tokens to get the tokens for an order, the resulting
list is ordered by token ID. Since we rely on the last added token being the one to be used
when renewing, we need to ensure that they are not ordered by ID. This commit changes the
get_order_tokens call to $renewal_order->get_payment_tokens to preserve the original order.

* Fix failed payment token update by saving token before order update

The woocommerce_subscriptions_changed_failing_payment_method_{gateway} hook is called when
the renewal order has its status changed. Therefore, we need to add the payment method to
the renewal order before completing the payment, otherwise, we cannot update the subscription
with the new payment details. This commit moves the token saving to before the order status
update and conciliates intents and setup intents to reuse checkout code.

* Prevent $0 payment notes from being added to orders

* Add comment about duplicate order payment tokens

* Check if token_id exists before calling WC_Payment_Tokens::get

This commit checks if a token_id is not false before fetching the token
to avoid unstable behavior. It also removes the empty array check, as
end() already returns false for empty arrays and adds logging in case
update_failing_payment_method cannot get a token.

* Remove unsupported subscription_payment_method_change_admin feature