Use current year + 1 as the test tokens expiry date
The test tokens are never really used, so we shoudln't have any issues if the
token expiry date has passed. However, if WC starts validating the expiry date
before creating an already expired token, they could start failing. This commit
uses the next year for the token expiration, to avoid this if it ever happens.