Add Saved Card component to display its data
This commit adds a Saved Card component which is used to display each card's data in
the user edit screen. This first version diverges and doesn't have some features from
figma, as its main use is to provide a way for merchants to check the token ID for a
saved payment method so they can edit a subscription's payment method.