Display token list when changing a subscription payment method
This commit removes the default input field with the payment method ID, for a select
element with all the user's tokens as options, making it easier for merchants to update
a subscription.

To do so, it uses a new script (subscription-edit-page.js) to make the DOM changes.
This script is localized with the subscriptions and tokens data and is compiled to
avoid issues with older browsers.