Use new Subscription action hook to admin payment meta inputs
This commit depends on the hook added in WC Subs #3798, which can
be used to provide a different meta input, rendering the JS from
subscription-edit-page.js unecessary. To support older Subscriptions
versions, the JS method was kept, but the plan is to remove it in the
future, when more merchants have migrated to a newer Subs version.