Docker Env: Removing the recursive `docker` directory from the plugin dir.
The whole docker environment is available in the plugin directory: `/var/www/html/wp-content/plugins/woocommerce-payments/docker`. This commit creates an empty volume and assigns it to the `woocommerce-payments/docker/` directory to make it empty inside the container.

It also creates another volume for `docker/bin`, which points to `/var/scripts` and is not accessible through the web.