Store test and live mode customer IDs in separate options
When a merchant toggles test mode on or off, the current customer saved
for a user is invalidated. This leads to that customer being recreated,
thus losing their saved methods and associated data on the platform.

This commit takes test mode into account when saving the customer ID
option, so a store can have both of those for the same user, preventing
errors when fetching a live customer in test mode and vice-versa.