Fix customer test mode when adding card from my account
When creating a setup intent, it is possible to get the same error we
get when fetching the payment methods for a customer. If the customer
belongs to the other mode than the current one, we need to migrate it
as well. This commit adds the check and migration when adding cards.