[#859] Introduced payments-specific exceptions and migrated from using the generic once.
While usage of generic exceptions is not the best practice, it was also suggested to not use them in [#826. Here we introducing several new exceptions and migrating existing codebase to use it:
- WC_Payments_AddPaymentMethod_Exception (payment creation behaviour)
- WC_Payments_InvalidPaymentMethod_Exception (payment method identification and usage behaviour)
- WC_Payments_ProcessPayment_Exception (payment processing behaviour)

This should enable us to provide more sophisticated orders handling (because you can catch specific exceptions and also handle the human-friendly error codes).