Install a specific version of xdebug (2.9.8) (#1173)
Running `pecl install xdebug` will now install version 3 of xdebug. This
breaks our existing configuration. We can fix it to 2.9.8 until we
update the config and confirm that the version of PHPUnit we're using
will work OK with version 3.

The main concern at the moment is difficulties we've had running unit
tests with coverage enabled in other repositories. Not something we're
doing here just yet, but I don't want to get into a state where we
can't.