Split code into multiple modules
Leaving everything in the DepositsList component file left everything
pretty cluttered. This splits all the JS imports into different files
into a structure we can follow with regular ReScript code going
forward.

The structure is pretty weird, but is made this way to avoid cyclic
dependencies. This way modules can reference each other using any of
the `Fully_Qualified_Module_Names`, but when general components are
using the modules they use the prettier `Fully.Qualified.Module.Names`.

This is necessary because of how the module system work. We could also
switch to a completely flat file structure, but I think we'd get naming
conflicts soon that way. We'll soon see that this structure is useful
once we start expanding on the data layer.