Update to add setting to enable/disable saving cards
This update adds a new setting which will enable/disable the ability to save cards on a site. If it is enabled, it will behave exactly as it currently does. If it is disabled, users will not be prompted to save their card for future purchases at the checkout, and will not be able to view/manage saved cards in their account.