Fix for issue with tokenization check
Fixes an issue where we would check to see whether saved cards were enabled before defining and initialising the settings. This could cause issues if a user updates and doesn't already have a saved_cards value enabled/disabled. Also updates some wording on the message that appears in the settings to explain the feature.