Update to add setting to enable/disable saving cards (#1180)
* Update to add setting to enable/disable saving cards

This update adds a new setting which will enable/disable the ability to save cards on a site. If it is enabled, it will behave exactly as it currently does. If it is disabled, users will not be prompted to save their card for future purchases at the checkout, and will not be able to view/manage saved cards in their account.

* Update changelog.txt and readme.txt

* Fix for some accidental re-formatting

* Fix for issue with tokenization check

Fixes an issue where we would check to see whether saved cards were enabled before defining and initialising the settings. This could cause issues if a user updates and doesn't already have a saved_cards value enabled/disabled. Also updates some wording on the message that appears in the settings to explain the feature.