Provide a better error message for minimum order amount requirement
WooCommerce payments might return an error on payment intent creation
if order amount is too low. This error will also contain a minimum order
amount value alongside with currency so client can present it in its own
way.

This change makes WooCommerce plugin to use this information to provide
a localized message to the end user.