Fix wcpaySettings is not defined on the settings page
With the introduction of zero decimal currency support, the
AccountFees component could not be rendered in the settings
page. That was caused due to `wcpaySettings` being added as
part of the `WCPAY_DASH_APP` script, which is only enqueued
on registered WC Admin pages.

This commit includes `wcpaySettings.zeroDecimalCurrencies`
as part of the `WCPAY_ADMIN_SETTINGS`, fixing the error.