Fix wp-mediaelement JS error
For whatever reason enabling typescript caused wp-medialement to be
enqueued at the wrong time, causing a JS error when WordPress tried to
use mediaelement.

We could've fixed this by adding a dependency on wp-mediaelement in PHP
manually when we register the WCPay Dashboard script, but I thought
adding wp-mediaelement as an external to the webpack config would fit
better, semantically, with our current setup.