Make exception message text-only.
`wc_price` tends to bloat resulting string with HTML tags and entities,
so to use it inside an exception message and other non-HTML places it
has to be stripped down.

To simplify the process `format_amount` method is added to
`WC_Payment_Utils` to apply proper amount transformations for amounts
used with Stripe and to return a properly formatted price without tags
and entities.