Fix missing token for free trial with saved card
We weren't saving the payment token to an order when the order includes
a free trial (or the first payment is $0 for a different reason) and
the subscription is paid for with a saved card.

This fixes the issue, and makes sure to save the token to the order
before returning early from the payment processing function.