Payment Request Button: Handle Puerto Rico country as United States (#1407)
* Map Puerto Rico country as UnitedStates

This is a port from a fix done in https://github.com/woocommerce/woocommerce-gateway-stripe/pull/1443

Puerto Rico is the only US territory that is supported by Stripe. Merchants with their country set to Puerto Rico (PR)  should be set as “US” on payment requests with Stripe.

* Make sure to get only one element as error message’s container

CSS selector is a bit generic and it’s present in other elements on the Product page resulting in the error message to be added multiple times.

To avoid breaking anything else we are not changing the selctor but making sure we get only the first one.

* Display error message on checkout page

* Simplify handling of error message

Also checks if the element exists before adding the error message

* Adjust CSS selector

Do not rely on theme sepcific classnames. Here we are chosing an element that will be always available for WooCommerce.

No longer need to expose `is_checkout_page` to the frontend.

* No need to expose is_checkout_page value to the frontend

* Run Prettier

* Place error message inside WooCommerce’s notices wrapper