Wrap wcpay enabled logic with feature flag
WooCommerce Payments + credit card are considered 1 payment method and when enabled,
credit card is enabled.

For backward compatibility, only when sepa/sofort/giropay feature is enabled, we will
change this logic so that we consider "WooCommerce Payments" and "Credit card" are
separate payment methods; each of these can be enabled/disabled separately.