Update deposit overview details (#1844)
* Override some wc summary component styles globally

We will be using the summary component outside of overview. So let's move the overrides to the main style.css

* Update deposit detail to add more overview details

Added fee percentage & amount and net amount. And differentiate instant deposits. Used a custom SummaryNumber component to avoid recreating the same styles, and share the same than the overview list.

* Update test and snapshot

* Add changelog entry

* Fix deposit gross and net calculation