Calculate coupon prices without adjustments
Coupon prices do not work well with adjustments and could cause
issues when rounding up by conceding more discount than intended
by the merchants. E.g. if the converted amount is 1 and rounding
is set to the tens, that amount would be ceiled to 10 and the
customer would get 10x the intended discount.