Add customer facing prices for customer multi-currency (#1860)
* Add type hint to get_default_currency

* Add Frontend_Prices to calculate prices for the frontend

Frontend_Prices currently supports simple and variation products,
shipping methods prices and taxes, and coupon amounts and min_amounts.
Also, it shouldn't be loaded on admin pages, modifying the prices
only for shoppers.

* Add support for setting that controls when charm should be applied

* Check if WC session exists before using it

Now that get_selected_currency is being used in the frontend,
it's possible that some request (e.g. an API request) doesn't
have a session set up. This commit checks if a session exists
before calling get/set to prevent fatal errors.

* Use ceil to the specified precision when rounding prices

* Add currency formatting to Frontend_Currencies

This commit uses the locale data from WooCommerce core (/i18n/locale-info.php)
and extracts the currency settings from it. It also adds the required methods
to properly format a currency in the frontend, dealing with the number of decimals,
decimal and thousand separators, and price formatting.

* Register frontend hooks with a priority higher than the default

* Calculate coupon prices without adjustments

Coupon prices do not work well with adjustments and could cause
issues when rounding up by conceding more discount than intended
by the merchants. E.g. if the converted amount is 1 and rounding
is set to the tens, that amount would be ceiled to 10 and the
customer would get 10x the intended discount.


* Run register_free_shipping_filters on init to prevent fatal

We cannot call WC_Shipping_Zones::get_zones() on plugin_loaded if the
store uses region-specific shipping zones, otherwise, WC throws a fatal.
This commit adds a filter to init, so register_free_shipping_filters
is called at an appropriate time.