TypeScript compatibility improvements (#1748)
 * Add TypeScript preset for babel to recognize TypeScript syntax
 * Update TypeScript settings and split babel & ts-loader tests
It looks like ts-loader and babel in one test set have a collision and
ts-loader starts reporting weird TypeScript errors like "Property has an
implicit any type".
 * Disable CSS class name linter checks for TypeScript too
 * Simplify set-up in favor of `ts-loader` and remove babel preset.
There are two options in linting TypeScript files. Since we already
have babel and ts-loader running simultaneously it is possible to remove
typescript preset for babel and reduce number of configurations to keep
an eye on.