Isolating `getDepositScheduleDescriptor` in utils.
Those tests used the whole `<DepositsOverview>` component to unit-test a specific function. Now that the function is used in two places, it makes sense to test it directly.

A schedule descriptor is still included in `<DepositsOverview />`, so it's test will make sure that it's included/rendered.

Note: Snapshots were not used for this, hence none are included in this commit.