Add getters for the Frontend Currencies and Prices instances
We're going to need to fetch some data from the conversion classes
when previewing the currency conversion in the settings pages.

This commit adds the getters for them in the Multi Currency class,
along with changing it so they're always instantiated. The filters
are still only registered with frontend requests.