Convert shipping amounts in _settings hook instead of woocommerce_package_rates
The woocommerce_package_rates approach has some incompatibilities with Subscriptions,
since it can be run more than once, resulting in successive conversions of the cost.

This commit modifies the existing free_shipping filters to be applied to all shipping
methods and also handle the cost conversion. As a result, we don't need to worry about
Subscriptions running the woocommerce_package_rates hook, and also don't have to
manually fix the taxes for the shipping packages as the correct value is used in the
calculation.