Update a copy of the $rate objects when converting shipping prices
Subscriptions caches the $rates array in woocommerce_package_rates early
and returns it re-applying the filter. If we modify the $rate object in
place, the conversion is applied more than once. This commit clones the
rate objects before modifying the cost and returns a new rates array to
prevent successive conversions in those cases.