Update a copy of the rate objects when converting shipping prices (#2146)
Subscriptions caches a reference to the $rates array early in
woocommerce_package_rates and returns it later, re-applying
that filter. If we modify the $rate object in place, the conversion
is applied more than once because the cached object will also
be updated during each conversion. This PR clones the rate
objects before modifying the cost and returns a new rates array
to prevent successive conversions in those cases.