Add get_customer_country method to Utils
We need a way to get the customer country to be able to use it
when picking a country-specific currency formatting and also
when automatically selecting a currency.

This currently doesn't take geolocation into account if there's
a customer object, so this will have to be extended later when
working on the currency selection.