Simplify add widget and refactor affected code
Use breadcrumb defaults to add the widget markup. To avoid the need of css code for each storefront child theme, making it easier to maintain.
Remove unneeded code, move the setting checkbox back to Settings class and update description to mention that the widget will be added to breadcrumb section.