Enable customer multi-currency for all merchants
This commit modifies the default value for the multi-currency
feature flag option to '1'. This results in the feature being
enabled for anyone that did not set it to be disabled.

This approach enables customer multi-currency for all merchants
and allows us to easily disable the entire extension if anything
goes wrong or if requested by some merchant.