Display Storefront Widget option when settings loaded with 1 currency
We currently override the settings to show the Storefront option on the
StorefrontIntegration class, but since that class is not instantiated
when enabled_currencies is 1 or less, the option is not displayed on the
first page load when there's only one currency enabled.

This commit moves the enabled currencies check before adding the widget
actions, instead of when instantiating the StorefrontIntegration class.