Remove e2e tests for deposits CSV export functionality
Turns out that the general consensus is to avoid e2e tests for non-critical
functionality in order to keep time to run the tests under check.

The tests weren't passing on GitHub as Stripe data is not synced during e2e setup.

We'll replace the removed e2e tests with unit tests.