Return upstream HTTP error code from terminal payment endpoint
Previously, we were returning a generic 502 HTTP status code for all
Stripe API errors. While 502 was not an inappropriate choice, it
represented *any* error at gateway or proxy, leading to the loss of the
actual status code returned by the upstream API.

For example, when we try to charge more than the amount in payment intent
Stripe fails with a 400 HTTP code which signifies that the input(s) received
by Stripe were incorrect or in bad format. Ignoring 400 and returning 502
meant that our endpoint's client had to rely on the error message to
determine the cause of error.