Check if is also an api request
This function was added to check if the request comes from an admin url to prevent converting currencies on admin analytics leaderboard. But this caused a new bug when you visit the store with an admin url as referer. Checking also if the request is an api request solves both issues.