Return upstream HTTP error code from terminal payment endpoint (#2884)
* Return upstream HTTP error code from terminal payment endpoint

Previously, we were returning a generic 502 HTTP status code for all
Stripe API errors. While 502 was not an inappropriate choice, it
represented *any* error at gateway or proxy, leading to the loss of the
actual status code returned by the upstream API.

For example, when we try to charge more than the amount in payment intent
Stripe fails with a 400 HTTP code which signifies that the input(s) received
by Stripe were incorrect or in bad format. Ignoring 400 and returning 502
meant that our endpoint's client had to rely on the error message to
determine the cause of error.

* Update tests for capture_charge to accommodate http code

* Add unit test for Stripe's 400 response scenario

As seen in issue #2390, Stripe returns this error message "Error: The
payment could not be captured because the requested capture amount is
greater than the authorized amount." with 400 as HTTP code.

* Add a code comment about $http_code

Clarify why $http_code doesn't need to be set inside $error_message's
"isn't empty" elseif block.

* Make minor adjustments to terminal payment invalid args test

* Rename the test function to a more generic name
* Make a stricter assertion with assertSame for HTTP code
* Add an assertion for error code
* Split the error message string into multiple lines

* Readjust the explicitly set HTTP codes

* Move the 200 assignment closer to where we handle API responses
* Replace 500 with 502 to represent unknown upstream errors
* Remove the now unnecesary code comment
* Refactor the unit tests as per this change