Add new method to retrieve enabled payment methods at checkout
The new get_upe_enabled_at_checkout_payment_method_ids method, as the name suggests, dinstincts from get_upe_enabled_payment_method_ids on the fact that is supposed to be used on checkout.

This distinction is needed because we need to include the capture rule that filters out payment methods that doesn't suport manual capture and it should be only present on checkout. It's also being added a check to see whether that mathod is enabled at checkout, which will be removed in the intent creation.