Trigger invoice payment after manual renewal
This triggers an invoice payment for migrated subscriptions on after
manual renewal is successful.

This also simplifies the invoice payment
cb to run on the `woocommerce_order_payment_status_changed` hook so we
don't have to worry about keeping track of order statuses. This hook is
only triggered when a order pending payment transitions to a paid
status.