Add check for invoice id before recording payment
This adds a check for order invoice id meta before recording payment to
distinguish between migration and regular renewals.

In the case of migration renewals, we want to mark the invoice as paid
out of band.

In the case of regular renewals, we want to ignore this since Stripe
initiated the renewal already.