Add tracking for enabling and disabling UPE (#2984)
* Trigger the events wcpay_upe_enabled/disabled when UPE is toggled on the client-side

This will allow tracking when UPE is being enabled and disabled on the client-side.

* Test tracking events being set for when enabling/disabling UPE

Testing that the tracking events wcpay_upe_enabled and wcpay_upe_disabled are set when toggling UPE.

* Track wcpay_upe_enabled when enabling UPE via the inbox note

When clicking on the inbox note for enabling new payment methods, UPE gets enabled. Track this toggling of UPE.

* Add unit tests for WC_Payments_Notes_Additional_Payment_Methods::maybe_enable_upe_feature_flag()

The existing tests only tested the early return for when the Stripe account isn't connected. Adding a test for the alternative flow when the account is connected.

* Check the UPE flag value changed before tracking events when toggled on the client side

It's unlikely but possible to trigger a 'wcpay_upe_enabled/disabled' event more than once without actually changing the current value. Avoid this by making sure the value changed before triggering the tracker event.

* Use an existing method for checking whether upe is enabled

This, instead of checking the option directly for conditionally trigger the track event for upe activation am deactivation.

* Trigger track events for wcpay_upe_enable/disable on upgrade under certain conditions

This event is only triggered once in existing installs to learn about the UPE status. It shouldn't run on new installs nor on installs were this initial status was already tracked.

* Remove tracking upe enabling/disabling on the server-side during REST API requests

WC_Tracker isn't loaded during REST API requests, so the event is never sent even though we're registering it.

* Track wcpay_upe_enabled/disabled on the client-side

Trigger the events wcpay_upe_enabled and wcpay_upe_disabled on the client-side when toggling UPE. We can't do this on the server-side with the current implementation because WC_Tracker isn't loaded during REST API requests.

* Remove unused 'use WCPay\Tracker' from the upe flag toggle controller

* Revert "Test tracking events being set for when enabling/disabling UPE"

This reverts commit de6145d85da1ba37d0624116f837ba7a05dc9a0f.

* Add tests for the Track_Upe_Status class

* Don't trigger the 'wcpay_upe_disabled' when upe has never been enabled on migration

During migration, we want to trigger the 'wcpay_upe_disabled' event only when UPE has been explicitly disabled by merchants. This commit updates the process so the event isn't triggered when the upe flag option isn't set.

* Record the 'wcpay_upe_enabled' directly when enabled via the admin notes

We were using Tracker::track_admin() before to do this, but because of the redirect the method has, the events are never recorded.

* Remove inaccurate test for tracking 'wcpay_upe_enabled' via admin notes

We can't check whether the event is registered because now we're recording this event directly.

* Remove leftover tests for upe_enabled/disabled track events for the inbox notice

This is recorded directly instead of using Tracker::track_admin(), so this test isn't relevant anymore.

* Remove the check to avoid redirecting during tests on the UPE enable note

It isn't bringing problems into the application, but it might be making some e2e tests fail. We're not testing this method with phpunit anymore anyway.

* Remove redundant check avoid tracking UPE in new installs

UPE track will be skipped because the UPE option won't be set in new installs, so this isn't needed.

* Add changelog entry for tracking UPE

* Remove inaccurate phpunit test for UPE tracking on new installs

* Move changelog entry from 3.1.0 to 3.2.0