Overwrite Gutenberg styles for the .components-visually-hidden element
Latest versions of Gutenberg overwrite the default wordpress' styles making the .components-visually-hidden element visible, thus breaking the UI. This commit adds the we need styles that are removed by Gutenberg, which is more a patch than a solution. This gets solved by upgrading @woocommerce/components to its latest version (17.0.0 as of today).