Refactor product service classes to accept a test mode variable
When a WC Product is updated, we need to update any test and live WCPay products that are linked to the
WC Product. To assist with this, we need to often get/set product and price meta data for both live and test environments, therefore
this commit updates all our helpers to accept a test mode param.