Update both test and live wcpay products when a wc product is updated
To avoid syncing issues that may occur when a store manager switches from live to test mode,
changes a product name and/or price, then switches back to live mode, the live product won't have any of the changes made to the WC Product while
the gateway was set in test mode.

To get around this, I've made it so our update product function will update both test and live WCPay products whenever a WC Product is updated.