Attach the payment intent data to renewal orders (#3113)
* Attach the payment intent data to renewal orders

This links the transaction with its order, and adds a note to the order referencing its transaction.

* Move get_and_attach_intent_info_to_order() from WC_Payments_Subscription_Service to WC_Payments_Invoice_Service

As mentioned in the PR, it makes more sense for this method to live in the Invoice service class than in the Subscription one.

* Rename parameter in the WC_Payments_Subscriptions's init() method

This variable could reference either the UPE or the card gateway. Changing its name to avoid confusion.

* Add newly introduced parameters and properties in php tests

WC_Payments_Invoice_Service new requires WC_Payment_Gateway_WCPay in its constructor. Adding it in the tests, as well as the payment_intent expected from the response's body when an invoice is paid, so they don't error out.

* Add tests for the newly introduced method

* Remove comma after the last argument in a constructor call

This throws an error in PHP < 7.3, thus removing it.