Fix single currency settings rounding and preview (#3127)
* Handle `none` option right

We were using the double pipe operator while checking the rounding, but as it comes as `0` instead of `none` was causing the issue. We need to switch it to a null coalescing operator to make it work properly.

* Handle `0` properly on preview calculation

Like we can receive `0` or `none` it's better to check if the parsed rounding float value is truthy

* Replace `none` with `0`

Get rid of the `none` keyword and use zero instead. We can even call `ceil_price` always, because the result of floatval would be `0.00` for every non number value.

* Replace `none` with `0` on the client too

* Add changelog entries