Reverse-engineer tax rate % figures for compounding rates
Stripe doesn't have a concept of applying rates in a compounding way. Therefore for us to apply
the correct tax amount, we need to reverse engineer the rate based on the amount of tax that applies in WC
so it can be reflected in Stripe billing.

The result of that essentially means a 5% tax rate is actually 5.1234% in the way it is applied.