Fix UPE add-payment validations
When UPE is enabled, adding a payment method with incomplete fields does
not trigger validation. This is because return_url is a required field
and we don't provide it (correctly so, as we just need to trigger
validation). This sets redirect option so that it only redirect when payment method
requires it. However, as this call will always contain incomplete
fields, redirection will never happen.