Don't init Subscriptions Core if WooCommerce\Admin\PluginsHelper is missing
When a store deactivates WooCommerce, WC Payments still tries to check if we need to init
Subscriptions Core. This causes a fatal error because `WooCommerce\Admin\pluginsHelper` doesn't exist.

In this commit, I've added a check to make sure we only use PluginsHelper and init subscriptions core if that class exists.