Limit period intervals in yearly subscription products (#3267)
* Force subscriptions period intervals to 1 for yearly periods

Stripe doesn't allow having period intervals greater than 1 year for subscriptions. This commit forces the interval to be '1' when the period is set to 'year'. This way we prevent the checkout processing from failing for shoppers when they purchase a subscription with an interval greater than 1 year.

* Add changelog entry

* Rename newly added methods to more concise names

* Validate intervals for daily, weekly, and monthly periods as well as yearly ones

Even though we have '6' as the maximum value for the interval options out of the box, third party developers can add greater values via hooks. We're also covering possible custom intervals that could have been added.

* Update the string in the error message displayed for invalid subscriptions intervals

We're now including the new billing cycle we've set for the subscription to replace the invalid one.

* Add different error messages to products and variations

 - Don't show multiple error messages for every variation that had issues. Just show one

Co-authored-by: mattallan <matt.allan@automattic.com>