Fix the link to customers table (#3315)
* Use reports customer id for admin/customers endpoint

WC-Admin analytics system uses wp_wc_customers_lookup table to report on all customers, including guest checkout which doesn't create an entry in wp_users.

The /customers path will always read from the wp_wc_customers_lookup table thus we need to provide id from that table instead of wp_users.

* Handle links to customers when analytics are disabled

If the WooCommerce Analytics feature is disabled in the setting we don't create an entry in the wp_wc_customers_lookup table. For checkouts made in this state, any links to customers are broken.

This commit ensures that we display customers as links only when we have an entry to display.