Display the UPE additional-payment-methods wizard even when the UPE preview is disabled
When the UPE preview is disabled, there shouldn't be a way to get to this page. But there are instances where the inbox note for setting up additional payment methods continues to show up when the preview option is updated in the database, as if it were cached. So I'm opting for displaying the wizard to prevent WCPay from looking broken if this inbox note is displayed in this case. Ideally, the inbox note wouldn't be displayed and there wouldn't be a way to get to the additional-payment-methods setup page. This is an edgy case, though. UPE preview is enabled by default.