Fix e2e tests related to the WooCommerce blocks' checkout (#3277)
* Fill up billing details instead of the shipping ones for WCB checkout e2e tests

We were filling up the shipping details for the WooCommerce blocks checkout. In the test setup, the billing details are asked instead of the shipping one. Because of this, the checkout process gets stuck and the tests fail.

* Fill the billing state input in WCB fill billing details e2e flow

The billing state field of the WooCommerce blocks checkout form is empty by default. Not filling this field prevents the checkout process from moving forward, thus making the tests fail.

* Wait until the Place Order button isn't disabled before clicking on it

It takes a bit for the button to be enabled on the checkout page after filling the billing form. Wait until it's enabled to click on it and proceed with the tests.

* Wait for the Place Orden button to be clickable in card failure tests

After entering the billing details, it takes a bit for the button to be clickable. This commit waits for it before proceeding with the tests.

* Remove unneeded fill of a billing field

Looks like this isn't relevant.

* Set User Agent before creating the WCB Checkout page

Leaving the 'puppeteer-debug wp-e2e-test' string we were using prevents the 'Add block' button from being enabled, and the execution gets stuck when trying to add the Checkout block to the page.

* Add a real user agent to all headful tests

Instead of setting the user agent when going to the Edit page screen, we're setting it up for all tests as discussed in the PR.

Co-authored-by: Achyuth Ajoy <achyuthajoy@gmail.com>
Co-authored-by: deepakpathania <68396823+deepakpathania@users.noreply.github.com>