Add redirect page to view a Capital loan offer
Merchants should have a way to view a Capital loan offer
from the link in the offer email. This commit adds such
a handler when an admin request is made with the wcpay-
view-capital-offer GET parameter.

The redirect will only happen if the request is not AJAX,
the user has admin permissions and the GET parameter is
specified. If we're not able to get an account link for
a merchant, they will be redirected to the Overview page.

Note: We don't check for a nonce here because it's an
automatic redirect. Navigating to the loan offer page is
also not going to accept the loan or take any actions for
the customer. The main risk we have here regarding CSRF
or clickjacking attacks is having merchants be redirected
to the Stripe's dashboard or the WC Pay Overview page.