Add redirect to view Capital financing offers (#3673)
* Add method to get a capital link to API client

Capital links are going to be used to display loan offers
to merchants. They work similarly to the login link, but
instead of the main dashboard, merchants are redirected
to a page to check the loan offer.

* Add redirect page to view a Capital loan offer

Merchants should have a way to view a Capital loan offer
from the link in the offer email. This commit adds such
a handler when an admin request is made with the wcpay-
loan-offer GET parameter.

The redirect will only happen if the request is not AJAX,
the user has admin permissions and the GET parameter is
specified. If we're not able to get an account link for
a merchant, they will be redirected to the Overview page.

Note: We don't check for a nonce here because it's an
automatic redirect. Navigating to the loan offer page is
also not going to accept the loan or take any actions for
the customer. The main risk we have here regarding CSRF
or clickjacking attacks is having merchants be redirected
to Stripe's dashboard or the WC Pay Overview page.

* Render an error message when loan offer link fails

Instead of just redirecting to the Overview page, the client will
now display an error message that the loan might have expired so
merchants are aware of the error.