Save email in classic checkout (#3824)
* Save email in classic checkout

A new field was added to move the email address to the top of checkout. The email wasn't being saved to the order. This updates the order meta with the posted email address.

* Filter core email field rather than unset it.

Unsetting the email field led to issues with the email address being saved to the order. This change filters the checkout fields, setting the core email field to an empty string, allowing WooCommerce to pull data from the new field of the same name.

* Update DocBlock