Be more defensive about expected elements on screen.
Users could have removed the `billing_phone` from their checkout form - account for that.

Users could have filtered the submit button html so don't rely on the submit button ID being the same as it is in core. Use `querySelector` to get the `button` with `type` `submit` inside `form.woocommerce-checkout`. Also check that we were able to select the element before attempting to alter the attributes.