Update email validation that we use before checking if the email exists in platform checkout.
Checks that the email address has at least a 2 character TLD. 1 character TLDs are technically valid but aren't used in practice. They also fail `is_email()` from WordPress which is what we use to validate emails on the platform checkout.

Also needed to disable eslint with the `/* eslint-disable */` because the linter was trying to format the line to fit in max length (even with `// eslint-disable-line max-len`) and I couldn't commit the change. If someone knows how I can better solve this, let me know.