Handle get document errors better
This commit refactors the API Client request method to handle
responses that are not JSON, and adds a property to return a
raw response, instead of just the body. This change is needed
because document requests rely on the headers and body to
properly forward the document to the merchant's browser.

It also handles any exceptions thrown by the get_document
method in the documents controller and presents the user with
a wp_die screen.